backend/requirements.txt
fastapi>=0.100.0 # Use a recent version
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
databases[sqlite]>=0.7.0
pydantic[email]>=2.0.0
python-dotenv>=0.20.0 # For loading .env files (optional but recommended)
python-multipart>=0.0.5 # Often needed for form data, good to include
aiosqlite>=0.17.0 # Explicitly add if needed, though databases[sqlite] should pull it in
asyncpg>=0.25.0 # Add async driver for PostgreSQL
psycopg2-binary>=2.9.0 # Add sync driver for PostgreSQL (needed for sync engine)